refactor: primitives handlers
